move generateExplorePathOfRotation to its own file
